backend/requirements.txt
Django
Unipath
dj-database-url
dj-static
gunicorn
django-decouple
mock
djangorestframework
markdown
django-filter
psycopg2
ipdb
django-extensions